==> 01-Program-Years/2021/Project Code/dspg21RnD/src/03_ai_wiki/01_scrape_wiki.ipynb <==
# Scrape AI Wiki

Author: Crystal

Date: June 21, 2021

import pandas as pd
import numpy as np
import json
import glob
import time


# web scrapping
import requests as r
from bs4 import BeautifulSoup
import re
import bs4 as bs
import urllib

import nltk


# Scrapping all the text

TODO: Not sure what's the difference between 'find_all' and 'select' functions from Beautiful Soup package. But it works.

url = 'https://en.wikipedia.org/wiki/Artificial_intelligence'
source = urllib.request.urlopen(url).read()

#parse the data and create beautifulsoup object
soup = BeautifulSoup(source, 'html.parser')

# scrape all the text from wiki
texts = ""
#for paragraph in soup.find_all('p'):
#    texts += paragraph.text
    
for paragraph in soup.select('p'):
    texts += paragraph.text

## Clean text and save as a text file

### Functions to clean text data

# check if a text is empty
def is_empty(texts):
    if (texts =='\n') or (texts=='\r'):
        return True
    else:
        return False

#clen wiki text
def clean_wiki_text(texts):
    if is_empty(texts):
        return None
    else:
        texts = re.sub(r'\[[0-9]*\]', ' ', texts) #removes numbers
       # texts = re.sub(r'\s+',' ',texts) #remove long blank space
       # texts = texts.lower() #lowercase text
       # texts = re.sub(r'\d', ' ', texts) 
       # texts = re.sub(r'\s+', ' ', texts)
        return(texts)

#wiki_texts_clean = clean_wiki_text(texts)
#print(wiki_texts_clean[1:1000])
#stopwords = stopwords.words('english')

#example 
t = "AI is ....[29]"
re.sub(r'\[[0-9]*\]', ' ', t)

#### Save cleaned file

#with open("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text"+".txt", "w", encoding="utf-8") as f:
#    for i in soup.select('p'):
 #       text_i = i.text
 #       text_i = clean_wiki_text(text_i)
 #       # write each paragraph to the file
 #       f.write(str(text_i))

# Scrapping table of contents

# nested items listed in the table of contents
topics = []
for topic in soup.find_all("h3"):
    topics.append(topic.text)
print(topics)

# Scrapping all the links

1. Exclude url links that contain:
    - ".png"
    - ".jpg"
    - ".svg"
    - ".gif"
    - "Template"
    - "(identifier)", might be a link to a journal article


2. Exclude duplicate links

url_ls = []
title_ls = []

condition = url.startswith("/wiki/") & (not url.endswith(".jpg"))  & (not url.endswith(".png"))  & (not "Template" in url) & (not url.endswith(".gif"))  & (not "(identifier)" in url)

for link in soup.find_all("a"):
    url = link.get("href", "")
    if url.startswith("/wiki/") & (not url.endswith(".jpg"))  & (not url.endswith(".png"))  & (not "Template" in url) & (not url.endswith(".gif"))  & (not "(identifier)" in url):
       # print(url)
        url_ls.append(url)
        title_ls.append(link.text.strip())


df = pd.DataFrame({'url':url_ls, 'title':title_ls})
df = df.drop_duplicates(subset = 'url', keep = 'first')

print(df.shape)
df.to_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_links.csv', index = False)   

Todo: 
1. Remove unrelated items
    - people's names
2. What to do with the topics listed under "Articles related to Artificial intelligence"?
3. What to do with acronames such as "k-NN"?

# Scrapping topics within the itemized list

Problem: suppose to grab three sections with class sidebar-content hlist, but only grabbing the first one


## I. AI, ML/data mining

# getting the headding too
div = soup.find('div', {"class": "sidebar-list-content mw-collapsible-content"})
topics = []
for topic in div.find_all("a"):
    topics.append(topic.text)
print(topics)

# Problem: suppose to grab all sections with class sidebar-content, but only grabbing the first one
# getting the items only (not heading)
td = soup.find('td', {"class": "sidebar-content"})
topics = []
for topic in td.find_all("a"):
    topics.append(topic.text)
print(topics)

## II. Evolutionary algorithm

# part of a series on evolutionary algorithm
td = soup.find('td', {"class": "sidebar-content hlist"})
topics = []
for topic in td.find_all("a"):
    topics.append(topic.text)
print(topics)
