==> 01-Program-Years/2021/Project Code/dspg21RnD/src/03_ai_wiki/02_ner_wiki_text.ipynb <==
# Named Entity Recognition on AI Wiki text

Author: Crystal

Major package used: Stanza 

import pandas as pd
import numpy as np
import json
import glob
import time


# web scrapping
import requests as r
from bs4 import BeautifulSoup
import re
import bs4 as bs
import urllib

# Gensim
import gensim
import gensim.corpora as corpora
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel

#name entity recognition
import stanza
stanza.download('en') # download English model

#spacy
import spacy
import nltk
from nltk.corpus import stopwords
#spacy.load("en_core_web_sm")

#visualization
import pyLDAvis
import pyLDAvis.gensim_models as gensimvis

import warnings
#warnings.simplefilter('always')
warnings.filterwarnings("ignore", category=DeprecationWarning)

## Read in AI Wiki text

#texts = open("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text.txt", "r")
with open("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text.txt") as f:
    contents = f.read()

len(contents)
print("First paragraph:",contents[1:1000])

print("Last paragraph:", contents[61100:61759])

type(contents)

## Name entity recognition

nlp = stanza.Pipeline(lang='en', processors='tokenize,ner') # initialize English neural pipeline, tokenize and named entity recognition

#Document will contain a list of Sentences, and the Sentences will contain lists of Tokens. 
doc = nlp(contents)

#extract named entities
entities = doc.entities

len(entities)

type(entities)

entities[0]

df = pd.DataFrame()
for entity in entities:
    df = df.append({
        "text": entity.text,
        'type': entity.type,
        'start_char': entity.start_char,
        'end_char': entity.end_char
        }, ignore_index = True)

### Save name entity recognition on AI Wiki text as a panda dataframe

#df.to_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text_entity.csv', index = True) #export csv
