==> 01-Program-Years/2021/Project Code/dspg21RnD/src/03_ai_wiki/04_clean_links.ipynb <==
# Clean AI Wiki Links

import pandas as pd
import numpy as np
import json
import glob
import time


# web scrapping
import requests as r
from bs4 import BeautifulSoup
import re
import bs4 as bs
import urllib

# Gensim
import gensim
import gensim.corpora as corpora
from gensim.utils import simple_preprocess
from gensim.models import CoherenceModel

#spacy
import spacy
import nltk
from nltk.corpus import stopwords

#visualization
import pyLDAvis
import pyLDAvis.gensim_models as gensimvis

import warnings
#warnings.simplefilter('always')
warnings.filterwarnings("ignore", category=DeprecationWarning)

## Read links data

links = pd.read_csv("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_links.csv") #import csv
links = links[links['title'].notnull()] #exclude rows that are na for title
links.title= links.title.str.lower() #lowercase title

links = links.rename(columns={'title': 'text'})
len(links)

links.head()

### De-duplicate Links

TODO: There are some same text but different links ones.

# no entry that have dupilicates on both url and text
links_dedup = links.drop_duplicates()
len(links_dedup)

## Read named entity data

named_entity = pd.read_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text_entity.csv') 
named_entity = named_entity[["text", "type"]]
named_entity.text= named_entity.text.str.lower() #lowercase title
named_entity = named_entity.rename(columns={'type': 'ner'})

named_entity

# Validate links 

## Approach I.If not exact match --> valid link

#Join two dataframes
links_ne = pd.merge(links,named_entity,on='text',how='left')
#links_ne.to_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_links_ner.csv', index = False)  
result_links_valid = links_ne[links_ne['ner'].isna()]
result_links_valid

### Examine NER in the links

links_matched = links_ne[links_ne['ner'].notnull()] #exclude rows that are na for title
ls_links_matched = links_matched.text.to_list()

def find_unique_in_list(ls):
    ls_unique = []
    for i in ls:
        if i not in ls_unique:
            ls_unique.append(i)
    return ls_unique
        
ls_links_matched_unique = find_unique_in_list(ls_links_matched)
ls_links_matched_unique

## Approach II.If NER does not contain topics(links) --> valid

ne_ls = named_entity["text"]
links_ls = links["text"]

links_valid_ls =[]
links_ner_ls = []
for link in links_ls:
    for ne in ne_ls:
        if link not in ne:
            links_valid_ls.append(link)
        elif link in ne:
            links_ner_ls.append(link)
links_valid_ls =  find_unique_in_list(links_valid_ls)
links_ner_ls=  find_unique_in_list(links_ner_ls)
print(len(links_valid_ls))
print(len(links_ner_ls))

#Links that are NER (invalid topics) 
links_ner_ls[0:50]

links_valid_ls[1:50]

links_valid = pd.DataFrame(links_valid_ls, columns =['text'])
links_valid = links_valid.assign(valid = True)

links_ner = pd.DataFrame(links_ner_ls, columns =['text'])
links_ner= links_ner.assign(valid = False)

valid_vs_ner = pd.concat([links_valid, links_ner])
    
result = pd.merge(links,valid_vs_ner,on='text',how='left')

result_links_valid = result[result['valid']==True]
result_links_valid

# Export valid links

result_links_valid.to_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_links_valid.csv', index = False)
