==> 01-Program-Years/2021/Project Code/dspg21RnD/src/03_ai_wiki/06_bert.ipynb <==
# Preliminary Sentence Bert Analysis

Author: Crystal

#bert
from sentence_transformers import SentenceTransformer, util
import torch

import pandas as pd

import re

import nltk
#nltk.download() #input: punkt

from nltk import tokenize

#embeddings
embedder = SentenceTransformer('paraphrase-MiniLM-L6-v2')
#embedder = SentenceTransformer("allenai/scibert_scivocab_uncased")

## Functions

def get_corpus_embeddings(dir):
    with open(dir) as f:
        ai_text = f.read()
    ai_corpus = tokenize.sent_tokenize(ai_text) #sentence tokenization
    ai_embeddings = embedder.encode(ai_corpus, show_progress_bar=True) # embeddings
    return ai_embeddings


ai_embeddings = get_corpus_embeddings("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/ai_wiki_text.txt")

# k: number of similar sentences from AI corpus
# abstract: abstract from FEDERAL RePORTER
# print_result: if TRUE, print out the similar sentenses from AI corpus to each sentence in the abstract
def get_score(k, abstract, print_result = False):
    queries = tokenize.sent_tokenize(abstract) 

    # init a result list for scores
    result = []
    
    # Find the closest k sentences of the AI corpus for each query sentence (ML) based on cosine similarity
    top_k = min(k, len(ai_corpus))
    
    for query in queries: #compare each sentence in the abstract to the ai corpus
        query_embedding = embedder.encode(query, show_progress_bar=False) 
        
        # We use cosine-similarity and torch.topk to find the highest k scores
        cos_scores = util.pytorch_cos_sim(query_embedding, ai_embeddings)[0]
        
        top_results = torch.topk(cos_scores, k=top_k)   #get the top k scores
        result.append(top_results.values.tolist()) #unlist the top result list
        if print_result:
            print("\n\n======================\n\n")
            print("Query:", query)
            print("Results:", top_results)
            print("\nTop k=5 most similar sentences in corpus:")
            for score, idx in zip(top_results[0], top_results[1]):
                print(ai_corpus[idx], "(Score: {:.4f})".format(score))
    return result

## AI related articles and clean

# functions to clean text
def is_empty(texts):
    if (texts =='\n') or (texts=='\r'):
        return True
    else:
        return False

#clen wiki text
def clean_wiki_text(texts):
    if is_empty(texts):
        return None
    else:
        texts = re.sub(r'\[[0-9]*\]', ' ', texts) #removes numbers
        texts = re.sub(r'\s+',' ',texts) #remove long blank space
        return(texts)
    
articles = pd.read_csv("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/wiki_ai_related_articles.csv") #import csv

texts = articles["article"] #get the text

text_clean_ls = []
for text in texts:
    text_clean = clean_wiki_text(text)
    text_clean_ls.append(text_clean)

### Example 1. Abstract on "FROM ALPHAGO TO POWER SYSTEM ARTIFICIAL INTELLIGENCE"

abstracts = pd.read_pickle("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/smaller-final-dataset.pkl")

abstracts[abstracts["PROJECT_TITLE"] =="FROM ALPHAGO TO POWER SYSTEM ARTIFICIAL INTELLIGENCE"]

We feed in ML Wiki article as a query.
For each sentence in the ML Wiki article, we identify the top 5 most similar/closest sentences from AI Wiki article based on cosine similarity

ai_related_p = abstracts["ABSTRACT"][531850]
ai_related_p

get_score(5, ai_related_p, print_result = False)

## Example 2. Abstract on "STRUCTURE OF SIGNAL PEPTIDE PEPTIDASE"

abstract_text = abstracts["ABSTRACT"][0]#get the text
abstract_text

get_score(5, abstract_text, print_result = False)

get_score(queries, print_result=False)

### Example


text_Basilica = "Pope Julius' scheme for the grandest building in Christendom was the subject of a competition for which a number of entries remain intact in the Uffizi Gallery, Florence. It was the design of Donato Bramante that was selected, and for which the foundation stone was laid in 1506. This plan was in the form of an enormous Greek Cross with a dome inspired by that of the huge circular Roman temple, the Pantheon.[7] The main difference between Bramante's design and that of the Pantheon is that where the dome of the Pantheon is supported by a continuous wall, that of the new basilica was to be supported only on four large piers. This feature was maintained in the ultimate design. Bramante's dome was to be surmounted by a lantern with its own small dome but otherwise very similar in form to the Early Renaissance lantern of Florence Cathedral designed for Brunelleschi's dome by Michelozzo. Bramante had envisioned that the central dome would be surrounded by four lower domes at the diagonal axes. The equal chancel, nave and transept arms were each to be of two bays ending in an apse. At each corner of the building was to stand a tower, so that the overall plan was square, with the apses projecting at the cardinal points. Each apse had two large radial buttresses, which squared off its semi-circular shape."

queries = tokenize.sent_tokenize(text_Basilica)
print(queries[0:10])
print(len(queries))

get_score(4, queries, print_result = True)

result = []

# Find the closest k sentences of the AI corpus for each query sentence (ML) based on cosine similarity
top_k = min(k, len(ai_corpus))

query = queries[0]

query_embedding = embedder.encode(query, show_progress_bar=False) 

# We use cosine-similarity and torch.topk to find the highest k scores
cos_scores = util.pytorch_cos_sim(query_embedding, ai_embeddings)[0]

top_results = torch.topk(cos_scores, k=top_k)   #get the top k scores
result.append(top_results.values.tolist()) #unlist the top result list

print("\n\n======================\n\n")
print("Query:", query)
print("Results:", top_results)
print("\nTop k=5 most similar sentences in corpus:")
for score, idx in zip(top_results[0], top_results[1]):
    print(ai_corpus[idx], "(Score: {:.4f})".format(score))
