==> 01-Program-Years/2021/Project Code/dspg21RnD/src/03_ai_wiki/08_embedding_score_statistics_calculation.ipynb <==
# Score Distribution

import pandas as pd
import statistics #calculate mean and others
import matplotlib.pyplot as plt
from scipy import stats
from scipy.stats import skew

#import data
abstracts_score = pd.read_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/abstracts_embedding_score_all.csv')   
abstracts = pd.read_pickle("/home/zz3hs/git/dspg21RnD/data/dspg21RnD/smaller-final-dataset.pkl")  #use Python 3.8 environment
oecd_ai = pd.read_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/nterm_matched_ai_abstracts.csv')   


print(len(abstracts_score))
print(len(abstracts))

abstracts_score_cb = abstracts.merge(abstracts_score, left_on = "original index", right_on = "original index")
abstracts_all = abstracts_score_cb.merge(oecd_ai, left_on = "original index", right_on = "original index", how = "left")
abstracts_all

sentence_score_ls = abstracts_all["sentence_score"]

sentence_score_ls_float = []
for sentence_score in sentence_score_ls:
    sentence_score = str(sentence_score)[1:-1]
    sentence_score = sentence_score.split(",")
    item_float= []
    for item in sentence_score:
        item_float.append(float(item))
    sentence_score_ls_float.append(item_float)

    
abstracts_all["sentence_score_float"] = sentence_score_ls_float

sentence_score_ls = abstracts_all["sentence_score_float"]

mean_abstract_score= []
num_sentences = []
range_sentence_score = []
max_sentence_score = []
median_sentence_score = []
skewness_sentence_score = []
for sentence_score in sentence_score_ls:
    mean_abstract_score.append(statistics.mean(sentence_score))
    num_sentences.append(len(sentence_score))
    range_sentence_score.append(max(sentence_score)- min(sentence_score))
    max_sentence_score.append(max(sentence_score))
    median_sentence_score.append(statistics.median(sentence_score))
    skewness_sentence_score.append(stats.skew(sentence_score))
abstracts_all["mean_abstract_score"]=mean_abstract_score
abstracts_all["num_sentences"]=num_sentences
abstracts_all["range_sentence_score"]=range_sentence_score
abstracts_all["max_sentence_score"]=max_sentence_score
abstracts_all["median_sentence_score"]=median_sentence_score
abstracts_all["skewness_sentence_score"]=skewness_sentence_score

abstracts_all = abstracts_all.assign(IS_AI = lambda x: x.IS_AI.notnull())
abstracts_all = abstracts_all.assign(IS_AI_1 = lambda x: x.IS_AI_1.notnull())
abstracts_all = abstracts_all.assign(IS_AI_2 = lambda x: x.IS_AI_2.notnull())

#abstracts_all.to_csv(r'/home/zz3hs/git/dspg21RnD/data/dspg21RnD/abstracts_embedding_score_stats.csv', index = False)  


abstracts_all_score
